+++ modules/hardware/src/main/java/org/cachyos/controlcenter/modules/hardware/HardwareBackend.java
@@ -1,0 +1,5 @@
+package org.cachyos.controlcenter.modules.hardware;
+
+public interface HardwareBackend {
+  HardwareSnapshot inspect();
+}